pyproject.toml
[project]
name = "inventory-service"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "supabase>=2.0",
    "python-dotenv>=1",
]

[tool.uv]
package = false
